Symbols as constants (closes #150)
I'm not sure about this. I like the idea. Yet, the
implementations is not perfect.

For one, you just can't `for in` or `for of` over symbol-keys!
So I had to change that to use `Object.getOwnPropertySymbols`
and `forEach` over that. Not a big deal, though. Kinda clunky I
feel it is.

Another issue is that now for dynamically generated test names
instead of
'test FOO_BAR_BAZ 4 with state...'
we have
'test actionNames.fooBarBaz 4 with state...'
or something like that.

Please look at the tests outputs. You may wish to do that
locally because then you can use the `ava -s` option, for
sorted output (because tests run serially).
`npm-run ava -sv` is the complete command.